requirements.txt
google-adk==1.21.0
google-cloud-aiplatform[evaluation]==1.132.0
pandas==2.3.3
